+++ src/Parallelepiped.h
@@ -38,8 +38,8 @@ protected:
 	void moveToCenterCoords();
 	void moveFromCenterCoords();
 
-    void addToPointsCoordsMatrix(int numRow, int numCol, int addValue);
-    void mulToPointsCoordsMatrix(int numRow, int numCol, int mulValue);
+    void addToPointsCoordsMatrix(int numRow, int numCol, double addValue);
+    void mulToPointsCoordsMatrix(int numRow, int numCol, double mulValue);
 
     // fill pointArr, give indexes of point Surface from matrix of points in parallelepiped
     void fillArrayOfSurfacePoints(Point* pointArr, int index1, int index2, int index3, int index4);
